Create Prime number.java
A prime number is a natural number greater than 1 that has no divisors other than 1 and itself. In this project, we implement a Prime Number Checker algorithm that determines if a given number is prime or not. This tool can be used to validate prime numbers in mathematical computations, cryptography algorithms, and more.


Sample Input : 
23

Sample Output :
The given number is prime